Split the planner into two files for readability.
Also extracted compileObjects(...) for readability

+ documentation
