+++ Sources/PoieticFlows/Metamodel/FlowsModel.swift
@@ -15,8 +15,8 @@ extension Metamodel {
     /// queries that define the [Stock and Flow](https://en.wikipedia.org/wiki/Stock_and_flow)
     /// model domain.
     ///
-    /// The basic object types are: ``/PoieticCore/ObjectType/Stock``, ``ObjectType/Flow``, ``ObjectType/Auxiliary``. More advanced
-    /// node type is ``/PoieticCore/ObjectType/GraphicalFunction``.
+    /// The basic object types are: `Stock`, `Flow`, `Auxiliary`. More advanced
+    /// node type is `GraphicalFunction`.
     ///
     /// - SeeAlso: `Metamodel` protocol description for more information and reasons
     /// behind this approach of describing the metamodel.
